+++ RELEASENOTES.md
@@ -8,11 +8,14 @@ __Breaking Changes__:
 
 - `torchvision.dataset.MNIST` will try more mirrors.
     - The thrown exception might be changed when it fails to download `MNIST`, `FashionMNIST` or `KMNIST`.
+- `ObjectDisposedException` will now be thrown when trying to use the disposed dispose scopes.
+    - The constructor of dispose scopes is no longer `public`. Use `torch.NewDisposeScope` instead.
 
 __API Changes__:
 
 - #1291 `Tensor.grad()` and `Tensor.set_grad()` have been replaced by a new property `Tensor.grad`.
     - A potential memory leak caused by `set_grad` has been resolved.
+- `Include` method of dispose scopes has been removed. Use `Attach` instead.
 
 __Bug Fixes__: